added compliment component and called compliment api
