+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
+import "./assets/index.css";
+import App from "./components/App.jsx";
 
 ReactDOM.render(
   <React.StrictMode>
